test: extract shared fake-vault helper, extend obsidian mock with rename/createBinary
