+++ src/utils/food-log/__tests__/food-log-storage.test.ts
@@ -1,39 +1,7 @@
-import { TFile, TFolder, Vault } from 'obsidian';
 import { readFoodLog, writeFoodLog, getFoodLogFilePath, ensureFoodLogDirectoryExists } from '../food-log-storage';
 import { FoodLog } from '../../../types/nutrition';
 import { PluginSettings, DEFAULT_SETTINGS } from '../../../types/settings';
-
-function createFakeVault(initialFiles: Record<string, string> = {}) {
-  const files = new Map(Object.entries(initialFiles));
-  const folders = new Set<string>();
-
-  const vault: Vault = {
-    getAbstractFileByPath: jest.fn((path: string) => {
-      if (files.has(path)) return new TFile(path);
-      if (folders.has(path)) return new TFolder(path);
-      return null;
-    }),
-    read: jest.fn(async (file: TFile) => files.get(file.path) ?? ''),
-    modify: jest.fn(async (file: TFile, data: string) => {
-      files.set(file.path, data);
-    }),
-    create: jest.fn(async (path: string, data: string) => {
-      files.set(path, data);
-      return new TFile(path);
-    }),
-    createFolder: jest.fn(async (path: string) => {
-      folders.add(path);
-      return new TFolder(path);
-    }),
-    process: jest.fn(async (file: TFile, fn: (data: string) => string) => {
-      const result = fn(files.get(file.path) ?? '');
-      files.set(file.path, result);
-      return result;
-    }),
-  };
-
-  return { vault, files, folders };
-}
+import { createFakeVault } from '../../../test-utils/fake-vault';
 
 const settings: PluginSettings = { ...DEFAULT_SETTINGS, logStoragePath: 'tracker/health/food/log' };
