+++ tsconfig.json
@@ -25,6 +25,7 @@
   "exclude": [
     "node_modules",
     "**/*.test.ts",
-    "__mocks__"
+    "__mocks__",
+    "src/test-utils"
   ]
 }
